refactor(logic): isolate SAT and SMT contracts
