--- src/jacobian/math/logic/_operations.py
@@ -1,1156 +1,0 @@
-"""Stateless logic operations with domain-owned contracts and kernels."""
-
-from __future__ import annotations
-
-import re
-import shutil
-import tempfile
-from enum import StrEnum
-from pathlib import Path
-from typing import Annotated, Literal, NamedTuple, Self
-
-from pydantic import (
-    Field,
-    StrictBool,
-    StrictInt,
-    field_validator,
-    model_validator,
-)
-from pydantic_core import PydanticCustomError
-
-from jacobian._models import StrictModel
-from jacobian.canonical import canonicalize_json
-from jacobian.catalog._examples import example
-from jacobian.catalog.models import MathTool
-from jacobian.math.logic._cnf import (
-    _MAX_VARIABLES,
-    CanonicalCnf,
-    CnfCanonicalizeRequest,
-    CnfCanonicalizeResult,
-    SatAssignmentCheckRequest,
-    SatAssignmentCheckResult,
-    canonicalize_cnf,
-    check_sat_assignment,
-)
-from jacobian.process import (
-    ProcessPlatformTools,
-    ProcessResourceLimits,
-    run_bounded_process,
-    worker_environment,
-)
-
-_MAX_SMTLIB_BYTES = 128_000
-# Structural SMT-LIB budgets, enforced before Z3 sees the source. Depth bounds
-# the recursive-descent parser's native stack per nesting level; compound terms
-# bound the assertion-DAG nodes the parser allocates and the solver preprocesses;
-# declarations bound the symbol table and the width of any returned model;
-# numeral digits bound the big-integer expansion of one literal spelling, the
-# quantity whose coefficient work measurably outgrows Z3's own timeout and
-# rlimit checks near 16k digits (4,096 keeps worst measured shapes within a
-# small multiple of the declared wall time).
-_MAX_SMTLIB_DEPTH = 512
-_MAX_SMTLIB_TERMS = 32_768
-_MAX_SMTLIB_DECLARATIONS = 4_096
-_MAX_SMTLIB_NUMERAL_DIGITS = 4_096
-# Request-scoped solver budgets beyond wall time. Z3 rlimit is a deterministic
-# work measure: identical requests cut off identically regardless of host load
-# or speed. The ceiling is orders of magnitude above admitted easy queries
-# (measured <1k units) while cutting runaway search within seconds at a
-# measured ~0.2-5M units/s across QF regimes. max_memory caps Z3's own arena so
-# exhaustion surfaces as typed UNKNOWN instead of host memory pressure.
-_SOLVER_RLIMIT = 20_000_000
-_SOLVER_MAX_MEMORY_MB = 1024
-_MAX_MODEL_BYTES = 64_000
-_MAX_LPR_STEPS = 2_048
-_MAX_LPR_CLAUSE_WIDTH = 128
-_MAX_LPR_HINT_IDS = 8_192
-_MAX_LPR_REPLAY_WORK = 2_000_000
-_LPR_WALL_SECONDS = 10
-_LPR_HEAP_MEBIBYTES = 64
-_LPR_STACK_MEBIBYTES = 16
-_LPR_ADDRESS_SPACE_BYTES = 128 * 1024 * 1024
-_LPR_PROCESS_OUTPUT_BYTES = 16_384
-_MAX_LPR_RESULT_BYTES = 9 * 1024 * 1024
-_CAKE_LPR_MANIFEST = Path("/usr/local/share/jacobian/cake-lpr.manifest")
-_CAKE_LPR_MANIFEST_CONTENT = (
-    "format=jacobian.cake-lpr/v1\n"
-    "upstream_commit=a36874a8b750b43fe4b385b8ddbf5b033e46a3fa\n"
-    "basis_ffi.c=8e30d84fdcb2177aa5571d7fa6661a2fae5ecfd56baa0ce49c65f9233a9f87cb\n"
-    "cake_lpr.S=2f3af32d55083839b3fa0e693afd817679c0b8944bef41def05a8b0ec72b7d4a\n"
-)
-_SUPPORTED_SMTLIB_COMMANDS = frozenset(
-    {"set-logic", "declare-const", "declare-fun", "assert", "check-sat"}
-)
-_UnknownResource = Literal["time", "work", "memory"]
-
-
-def _validation_error(code: str, message: str) -> PydanticCustomError:
-    return PydanticCustomError(code, message)
-
-
-class SatSolveRequest(StrictModel):
-    cnf: CanonicalCnf
-    timeout_ms: StrictInt = Field(default=1_000, ge=1, le=10_000)
-
-
-class SatSolveResult(StrictModel):
-    outcome: Literal["SAT", "UNSAT", "UNKNOWN"]
-    assignment: tuple[StrictBool, ...] | None = Field(
-        default=None, max_length=_MAX_VARIABLES
-    )
-    exhausted: _UnknownResource | None = Field(default=None)
-    detail: str | None = Field(default=None, max_length=1_024)
-
-    @model_validator(mode="after")
-    def bind_assignment_to_outcome(self) -> Self:
-        if (self.outcome == "SAT") != (self.assignment is not None):
-            raise _validation_error(
-                "logic.sat_assignment_outcome",
-                "only a SAT result may carry an assignment",
-            )
-        if self.exhausted is not None and self.outcome != "UNKNOWN":
-            raise _validation_error(
-                "logic.unknown_exhaustion",
-                "only an UNKNOWN result may name an exhausted budget",
-            )
-        return self
-
-
-class LprPropagationHint(StrictModel):
-    """One named LPR propagation check and its ordered unit hints."""
-
-    clause_id: StrictInt = Field(
-        ge=1,
-        description=(
-            "A currently live clause-ID label whose propagation is checked. "
-            "Labels may be sparse solver-assigned values."
-        ),
-    )
-    at_hint_clause_ids: tuple[StrictInt, ...] = Field(
-        max_length=_MAX_LPR_HINT_IDS,
-        description=(
-            "Ordered positive clause IDs used for the propagation check of this "
-            "named live clause."
-        ),
-    )
-
-    @field_validator("at_hint_clause_ids")
-    @classmethod
-    def require_positive_hint_ids(
-        cls, value: tuple[StrictInt, ...]
-    ) -> tuple[StrictInt, ...]:
-        if any(clause_id <= 0 for clause_id in value):
-            raise _validation_error(
-                "logic.lpr_hint_id", "LPR hint clause IDs must be positive"
-            )
-        return value
-
-
-class LprAddition(StrictModel):
-    """One source-bound ASCII LPR PR/RAT addition."""
-
-    kind: Literal["addition"] = "addition"
-    clause_id: StrictInt = Field(
-        ge=1,
-        description=(
-            "A fresh positive clause-ID label above all canonical source clause "
-            "numbers; it may be a sparse solver-assigned value and may not "
-            "overwrite a live clause."
-        ),
-    )
-    clause: tuple[StrictInt, ...] = Field(
-        max_length=_MAX_LPR_CLAUSE_WIDTH,
-        description="Ordered nonzero literals on the exact CNF variable axis.",
-    )
-    witness: tuple[StrictInt, ...] | None = Field(
-        default=None,
-        max_length=_MAX_LPR_CLAUSE_WIDTH,
-        description=(
-            "Optional ordered PR witness. When present, its first literal equals "
-            "the added clause's first literal."
-        ),
-    )
-    at_hint_clause_ids: tuple[StrictInt, ...] = Field(
-        max_length=_MAX_LPR_HINT_IDS,
-        description=(
-            "Ordered currently live clause IDs for the addition's "
-            "asymmetric-tautology check."
-        ),
-    )
-    propagation_hints: tuple[LprPropagationHint, ...] = Field(
-        max_length=_MAX_LPR_HINT_IDS,
-        description=(
-            "At most one propagation hint per currently live clause ID, in the "
-            "order written to the LPR proof."
-        ),
-    )
-
-    @field_validator("at_hint_clause_ids")
-    @classmethod
-    def require_positive_at_hint_ids(
-        cls, value: tuple[StrictInt, ...]
-    ) -> tuple[StrictInt, ...]:
-        if any(clause_id <= 0 for clause_id in value):
-            raise _validation_error(
-                "logic.lpr_hint_id", "LPR hint clause IDs must be positive"
-            )
-        return value
-
-    @model_validator(mode="after")
-    def bind_witness_to_the_clause_pivot(self) -> Self:
-        if self.witness is not None:
-            if not self.clause:
-                raise _validation_error(
-                    "logic.lpr_witness_clause",
-                    "an empty LPR clause may not carry a witness",
-                )
-            if not self.witness or self.witness[0] != self.clause[0]:
-                raise _validation_error(
-                    "logic.lpr_witness_pivot",
-                    "an LPR witness must start with the added clause's pivot literal",
-                )
-        return self
-
-
-class LprDeletion(StrictModel):
-    """One ASCII LPR deletion of currently live clause IDs."""
-
-    kind: Literal["deletion"] = "deletion"
-    clause_ids: tuple[StrictInt, ...] = Field(
-        min_length=1,
-        max_length=_MAX_LPR_HINT_IDS,
-        description="Distinct currently live clause IDs to remove before the next step.",
-    )
-
-    @field_validator("clause_ids")
-    @classmethod
-    def require_positive_distinct_clause_ids(
-        cls, value: tuple[StrictInt, ...]
-    ) -> tuple[StrictInt, ...]:
-        if any(clause_id <= 0 for clause_id in value):
-            raise _validation_error(
-                "logic.lpr_deletion_id", "deleted LPR clause IDs must be positive"
-            )
-        if len(set(value)) != len(value):
-            raise _validation_error(
-                "logic.lpr_duplicate_deletion",
-                "one LPR deletion may not name a clause more than once",
-            )
-        return value
-
-
-LprStep = Annotated[LprAddition | LprDeletion, Field(discriminator="kind")]
-
-
-class SatLprRefutation(StrictModel):
-    """A bounded typed LPR/ASCII-v1 refutation, without checker syntax or flags."""
-
-    profile: Literal["LPR_ASCII_V1"] = "LPR_ASCII_V1"
-    steps: tuple[LprStep, ...] = Field(
-        max_length=_MAX_LPR_STEPS,
-        description=(
-            "Ordered LPR additions and deletions. The checker uses canonical source "
-            "clause IDs 1..m; every hint and deletion must name a currently live ID. "
-            "The derived literal-inspection work must not exceed 2,000,000."
-        ),
-    )
-
-
-class SatRefutationCheckRequest(StrictModel):
-    cnf: CanonicalCnf
-    refutation: SatLprRefutation = Field(
-        description=(
-            "One source-bound LPR/ASCII-v1 derivation. It uses the CNF's exact "
-            "one-based canonical clause order and variable axis."
-        )
-    )
-
-    @model_validator(mode="after")
-    def require_source_bound_lpr_profile(self) -> Self:
-        _validate_lpr_refutation(self.cnf, self.refutation)
-        return self
-
-
-class SatRefutationCheckResult(StrictModel):
-    """A source-bound LPR replay outcome; only VALID_REFUTATION proves UNSAT."""
-
-    outcome: Literal[
-        "VALID_REFUTATION", "INVALID_REFUTATION", "UNAVAILABLE", "TIMEOUT", "ERROR"
-    ]
-    cnf: CanonicalCnf
-    refutation: SatLprRefutation
-    detail: str | None = Field(default=None, max_length=1_024)
-
-    @model_validator(mode="after")
-    def bind_execution_detail(self) -> Self:
-        if (self.outcome == "VALID_REFUTATION") != (self.detail is None):
-            raise _validation_error(
-                "logic.refutation_detail",
-                "only a valid refutation may omit its outcome detail",
-            )
-        return self
-
-
-def _require_live_lpr_ids(
-    clause_ids: tuple[int, ...],
-    live_clause_widths: dict[int, int],
-    label: str,
-) -> None:
-    missing = next(
-        (clause_id for clause_id in clause_ids if clause_id not in live_clause_widths),
-        None,
-    )
-    if missing is not None:
-        raise _validation_error(
-            "logic.lpr_live_clause", f"{label} references non-live clause ID {missing}"
-        )
-
-
-def _require_lpr_literal_axis(
-    literals: tuple[int, ...], variable_count: int, label: str
-) -> None:
-    if any(literal == 0 or abs(literal) > variable_count for literal in literals):
-        raise _validation_error(
-            "logic.lpr_literal_axis",
-            f"{label} literal is outside the CNF variable axis",
-        )
-
-
-def _lpr_addition_work(
-    step: LprAddition,
-    live_clause_widths: dict[int, int],
-) -> int:
-    candidate_width = len(step.clause) + len(step.witness or ())
-    inspection_factor = candidate_width + 1
-    total = sum(width + 1 for width in live_clause_widths.values()) * inspection_factor
-    total += sum(
-        (live_clause_widths[clause_id] + 1) * inspection_factor
-        for clause_id in step.at_hint_clause_ids
-    )
-    return total + sum(
-        (live_clause_widths[hint.clause_id] + 1) * inspection_factor
-        + sum(
-            (live_clause_widths[at_clause_id] + 1) * inspection_factor
-            for at_clause_id in hint.at_hint_clause_ids
-        )
-        for hint in step.propagation_hints
-    )
-
-
-def _validate_lpr_addition(
-    step: LprAddition,
-    *,
-    variable_count: int,
-    source_clause_count: int,
-    live_clause_widths: dict[int, int],
-) -> None:
-    if step.clause_id <= source_clause_count:
-        raise _validation_error(
-            "logic.lpr_clause_id",
-            "LPR additions must use IDs after the canonical source clauses",
-        )
-    if step.clause_id in live_clause_widths:
-        raise _validation_error(
-            "logic.lpr_duplicate_clause_id",
-            "LPR additions may not overwrite a live clause ID",
-        )
-    _require_lpr_literal_axis(step.clause, variable_count, "LPR clause")
-    if step.witness is not None:
-        _require_lpr_literal_axis(step.witness, variable_count, "LPR witness")
-    _require_live_lpr_ids(
-        step.at_hint_clause_ids, live_clause_widths, "LPR asymmetric-tautology hint"
-    )
-    if len({hint.clause_id for hint in step.propagation_hints}) != len(
-        step.propagation_hints
-    ):
-        raise _validation_error(
-            "logic.lpr_duplicate_hint_id",
-            "LPR propagation hint clause IDs must be unique",
-        )
-    for hint in step.propagation_hints:
-        _require_live_lpr_ids(
-            (hint.clause_id,), live_clause_widths, "LPR propagation hint"
-        )
-        _require_live_lpr_ids(
-            hint.at_hint_clause_ids,
-            live_clause_widths,
-            "LPR propagation asymmetric-tautology hint",
-        )
-
-
-def _validate_lpr_refutation(cnf: CanonicalCnf, refutation: SatLprRefutation) -> None:
-    live_clause_widths = {
-        index: len(clause) for index, clause in enumerate(cnf.clauses, 1)
-    }
-    source_clause_count = len(live_clause_widths)
-    total_work = 0
-    for step in refutation.steps:
-        if isinstance(step, LprDeletion):
-            _require_live_lpr_ids(step.clause_ids, live_clause_widths, "LPR deletion")
-            for clause_id in step.clause_ids:
-                del live_clause_widths[clause_id]
-            continue
-        _validate_lpr_addition(
-            step,
-            variable_count=len(cnf.variables),
-            source_clause_count=source_clause_count,
-            live_clause_widths=live_clause_widths,
-        )
-        total_work += _lpr_addition_work(step, live_clause_widths)
-        if total_work > _MAX_LPR_REPLAY_WORK:
-            raise _validation_error(
-                "logic.lpr_work_budget",
-                "LPR replay exceeds the declared literal-inspection work bound",
-            )
-        live_clause_widths[step.clause_id] = len(step.clause)
-    echoed_result = {
-        "outcome": "INVALID_REFUTATION",
-        "cnf": cnf.model_dump(mode="json"),
-        "refutation": refutation.model_dump(mode="json"),
-        "detail": "x" * 1_024,
-    }
-    if len(canonicalize_json(echoed_result)) > _MAX_LPR_RESULT_BYTES:
-        raise _validation_error(
-            "logic.lpr_result_budget",
-            "LPR refutation exceeds the source-bound result limit",
-        )
-
-
-class SmtLogic(StrEnum):
-    QF_UF = "QF_UF"
-    QF_LIA = "QF_LIA"
-    QF_LRA = "QF_LRA"
-
-
-def _consume_smtlib_string(source: str, position: int) -> int:
-    position += 1
-    while position < len(source):
-        if source[position] != '"':
-            position += 1
-        elif position + 1 < len(source) and source[position + 1] == '"':
-            position += 2
-        else:
-            return position + 1
-    return position
-
-
-def _consume_smtlib_quoted_symbol(source: str, position: int) -> int:
-    r"""Scan a quoted SMT-LIB symbol, handling escaped bars (\| inside |...|)."""
-    pos = position + 1
-    while pos < len(source):
-        if source[pos] == chr(92) and pos + 1 < len(source) and source[pos + 1] == "|":
-            pos += 2
-        elif source[pos] == "|":
-            return pos + 1
-        else:
-            pos += 1
-    return len(source)
-
-
-def _consume_smtlib_atom(source: str, position: int) -> int:
-    while (
-        position < len(source)
-        and not source[position].isspace()
-        and source[position] not in ';()"|'
-    ):
-        position += 1
-    return position
-
-
-def _tokenize_smtlib(source: str) -> tuple[str, ...]:
-    """Tokenize enough SMT-LIB syntax to distinguish real command forms."""
-
-    tokens: list[str] = []
-    position = 0
-    while position < len(source):
-        character = source[position]
-        if character.isspace():
-            position += 1
-        elif character == ";":
-            newline = source.find("\n", position)
-            position = len(source) if newline == -1 else newline + 1
-        elif character in "()":
-            tokens.append(character)
-            position += 1
-        elif character == '"':
-            position = _consume_smtlib_string(source, position)
-            tokens.append("<string>")
-        elif character == "|":
-            position = _consume_smtlib_quoted_symbol(source, position)
-            tokens.append("<quoted-symbol>")
-        else:
-            end = _consume_smtlib_atom(source, position)
-            tokens.append(source[position:end])
-            position = end
-    return tuple(tokens)
-
-
-def _top_level_smtlib_commands(source: str) -> tuple[tuple[str, ...], ...]:
-    """Return the heads and immediate atoms of complete top-level commands.
-
-    This deliberately does not parse SMT-LIB.  Z3 remains the parser and solver;
-    the small lexical pass only prevents comments, strings, and nested terms from
-    impersonating the two boundary commands whose presence this contract owns.
-    """
-
-    commands: list[tuple[str, ...]] = []
-    command: list[str] = []
-    depth = 0
-    for token in _tokenize_smtlib(source):
-        if token == "(":
-            if depth == 0:
-                command = []
-            depth += 1
-        elif token == ")":
-            depth -= 1
-            if depth == 0:
-                commands.append(tuple(command))
-        elif depth == 1:
-            command.append(token)
-    return tuple(commands)
-
-
-class _SmtLibStructure(NamedTuple):
-    """Lexical structure of one SMT-LIB source, measured without parsing it."""
-
-    max_depth: int
-    compound_terms: int
-    numeral_digits: int
-
-
-def _atom_numeral_weight(atom: str) -> int:
-    """Return the digit width of one classified numeric-literal token.
-
-    Numeral, decimal, and bit-vector spellings expand into big integers or
-    rationals inside the solver. Digits inside simple symbols are interned
-    names and carry no weight; malformed tokens remain the backend parser's
-    typed rejection.
-    """
-
-    if atom.startswith("#"):
-        return max(len(atom) - 2, 0)
-    if atom.isdigit():
-        return len(atom)
-    head, separator, tail = atom.partition(".")
-    if separator and head.isdigit() and (not tail or tail.isdigit()):
-        return len(head) + len(tail)
-    return 0
-
-
-def _smtlib_structure(source: str) -> _SmtLibStructure:
-    """Measure nesting depth, compound-term count, and numeral width lexically.
-
-    This is the same deliberately non-parsing scan used for command shape; it
-    bounds what ``z3.parse_smt2_string`` may build before that parser runs.
-    An indexed literal such as ``(_ bvN w)`` spells its value inside a simple
-    symbol, so index position decides whether ``bvN`` digits carry numeral
-    weight; elsewhere digits stay interned names.
-    """
-
-    max_depth = depth = compound_terms = numeral_digits = 0
-    indexed_depth: int | None = None
-    previous_token = ""
-    for token in _tokenize_smtlib(source):
-        if token == "(":
-            depth += 1
-            compound_terms += 1
-            if depth > max_depth:
-                max_depth = depth
-        elif token == ")":
-            if indexed_depth == depth:
-                indexed_depth = None
-            depth -= 1
-        elif previous_token == "(" and token == "_":
-            indexed_depth = depth
-        else:
-            weight = _atom_numeral_weight(token)
-            if (
-                indexed_depth is not None
-                and token.startswith("bv")
-                and token[2:].isdigit()
-            ):
-                weight = len(token) - 2
-            if weight > numeral_digits:
-                numeral_digits = weight
-        previous_token = token
-    return _SmtLibStructure(max_depth, compound_terms, numeral_digits)
-
-
-_Z3_SOURCE_DIAGNOSTIC = re.compile(r'\(error "line \d+ column \d+: ')
-
-
-def _exception_message(exc: Exception) -> str:
-    message = exc.args[0] if exc.args else ""
-    if isinstance(message, bytes):
-        return message.decode("ascii", errors="replace")
-    return str(message)
-
-
-def _is_smtlib_source_diagnostic(exc: Exception) -> bool:
-    """Report whether one backend parse exception diagnoses the caller's source.
-
-    Z3's SMT-LIB front end reports caller-correctable source problems as
-    ``(error "line L column C: <diagnostic>")``. Backend conditions such as
-    exhausted memory or interruption surface through Z3's fixed error-code
-    message table and carry no source locator. A located diagnostic names a
-    grammar defect regardless of which resource keywords its text contains,
-    because the diagnostic quotes caller-controlled source spellings that may
-    legitimately contain words such as ``memory``.
-    """
-
-    return _Z3_SOURCE_DIAGNOSTIC.search(_exception_message(exc)) is not None
-
-
-def _require_parseable_smtlib(source: str) -> None:
-    """Reject source that Z3's SMT-LIB 2 parser cannot read as a request error.
-
-    Admission runs the same non-evaluating backend parse that execution uses,
-    so a schema-admitted request never discovers malformed syntax through an
-    execution exception. Only located parser diagnostics establish malformed
-    input; resource or other backend failures carry no evidence about the
-    source, so they defer to execution, which reports them as typed UNKNOWN.
-    Backend absence here is left to execution, where it keeps its typed
-    initialization outcome.
-    """
-
-    try:
-        import z3  # type: ignore[import-untyped]
-    except (ImportError, OSError):
-        return
-    try:
-        z3.parse_smt2_string(source)
-    except (z3.Z3Exception, OSError) as exc:
-        if not _is_smtlib_source_diagnostic(exc):
-            return
-        raise _validation_error(
-            "logic.smtlib_grammar",
-            "SMT-LIB input could not be parsed by the declared logic",
-        ) from exc
-
-
-class SmtSolveRequest(StrictModel):
-    logic: SmtLogic
-    smtlib: str = Field(
-        min_length=1,
-        max_length=_MAX_SMTLIB_BYTES,
-        description=(
-            "ASCII SMT-LIB that declares logic, contains exactly one check-sat command, "
-            "and ends with that command. Bounded before parsing: nesting depth at most "
-            f"{_MAX_SMTLIB_DEPTH}, compound terms at most {_MAX_SMTLIB_TERMS}, declared "
-            f"symbols at most {_MAX_SMTLIB_DECLARATIONS}, and any one numeral, decimal, "
-            f"or indexed bit-vector spelling at most {_MAX_SMTLIB_NUMERAL_DIGITS} digits. "
-            "The source must be well-formed SMT-LIB 2 for the declared logic; malformed "
-            "input is rejected during request validation."
-        ),
-        examples=[
-            "(set-logic QF_LIA)\n(declare-const x Int)\n(assert (> x 0))\n(check-sat)"
-        ],
-    )
-    timeout_ms: StrictInt = Field(default=1_000, ge=1, le=10_000)
-
-    @model_validator(mode="after")
-    def require_single_smtlib_query(self) -> Self:
-        try:
-            encoded = self.smtlib.encode("ascii")
-        except UnicodeEncodeError as exc:
-            raise _validation_error(
-                "logic.smtlib_ascii", "SMT-LIB input must be ASCII"
-            ) from exc
-        if len(encoded) > _MAX_SMTLIB_BYTES:
-            raise _validation_error(
-                "logic.smtlib_byte_budget", "SMT-LIB input exceeds the byte limit"
-            )
-        structure = _smtlib_structure(self.smtlib)
-        if structure.max_depth > _MAX_SMTLIB_DEPTH:
-            raise _validation_error(
-                "logic.smtlib_depth_budget",
-                f"SMT-LIB nesting exceeds the maximum term depth of {_MAX_SMTLIB_DEPTH}",
-            )
-        if structure.compound_terms > _MAX_SMTLIB_TERMS:
-            raise _validation_error(
-                "logic.smtlib_term_budget",
-                f"SMT-LIB exceeds the maximum of {_MAX_SMTLIB_TERMS} compound terms",
-            )
-        if structure.numeral_digits > _MAX_SMTLIB_NUMERAL_DIGITS:
-            raise _validation_error(
-                "logic.smtlib_numeral_budget",
-                "SMT-LIB contains a numeral wider than "
-                f"{_MAX_SMTLIB_NUMERAL_DIGITS} digits",
-            )
-        commands = _top_level_smtlib_commands(self.smtlib)
-        declarations = sum(
-            1
-            for command in commands
-            if command[:1] in (("declare-const",), ("declare-fun",))
-        )
-        if declarations > _MAX_SMTLIB_DECLARATIONS:
-            raise _validation_error(
-                "logic.smtlib_declaration_budget",
-                f"SMT-LIB declares more than {_MAX_SMTLIB_DECLARATIONS} symbols",
-            )
-        logic_commands = tuple(
-            command for command in commands if command[:1] == ("set-logic",)
-        )
-        if logic_commands != (("set-logic", self.logic.value),):
-            raise _validation_error(
-                "logic.smtlib_logic_declaration",
-                "SMT-LIB input must declare the requested logic",
-            )
-        if commands.count(("check-sat",)) != 1:
-            raise _validation_error(
-                "logic.smtlib_check_sat_count",
-                "SMT-LIB input must contain exactly one check-sat command",
-            )
-        if commands[-1:] != (("check-sat",),):
-            raise _validation_error(
-                "logic.smtlib_check_sat_position",
-                "SMT-LIB input must end with its check-sat command",
-            )
-        for command in commands:
-            if command and command[0] not in _SUPPORTED_SMTLIB_COMMANDS:
-                raise _validation_error(
-                    "logic.smtlib_command", f"unsupported SMT-LIB command: {command[0]}"
-                )
-        self._complete_backend_admission()
-        return self
-
-    def _complete_backend_admission(self) -> None:
-        """Parse through the backend once every structural check has passed.
-
-        Called at the end of ``require_single_smtlib_query``. A subclass whose
-        own ``mode="after"`` validators impose a stricter envelope overrides
-        this hook so out-of-envelope source never reaches the backend parser;
-        the most-derived request completes backend admission.
-        """
-
-        _require_parseable_smtlib(self.smtlib)
-
-
-class SmtSolveResult(StrictModel):
-    outcome: Literal["SAT", "UNSAT", "UNKNOWN"]
-    model_smtlib: str | None = Field(default=None, max_length=_MAX_MODEL_BYTES)
-    exhausted: _UnknownResource | None = Field(default=None)
-    detail: str | None = Field(default=None, max_length=1_024)
-
-    @model_validator(mode="after")
-    def bind_model_to_outcome(self) -> Self:
-        if (self.outcome == "SAT") != (self.model_smtlib is not None):
-            raise _validation_error(
-                "logic.sat_model_outcome", "only a SAT result may carry a model"
-            )
-        if self.exhausted is not None and self.outcome != "UNKNOWN":
-            raise _validation_error(
-                "logic.unknown_exhaustion",
-                "only an UNKNOWN result may name an exhausted budget",
-            )
-        return self
-
-
-_EXHAUSTION_DETAILS: dict[_UnknownResource, str] = {
-    "work": "the bounded solver work budget was exhausted",
-    "memory": "the bounded solver memory budget was exhausted",
-    "time": "the bounded solver time budget was exhausted",
-}
-
-
-def _classify_exhaustion(message: str) -> _UnknownResource | None:
-    """Classify one Z3 reason or exception message onto the exhausted budgets.
-
-    Exhaustion keywords classify only backend conditions, which carry no
-    source locator. A message containing a located ``(error "line ...
-    column ...: ...")`` diagnostic is never classified as exhaustion, even
-    when its text mentions a resource keyword: the diagnostic quotes
-    caller-controlled source spellings, so an undeclared identifier named
-    ``memory`` or a comment mentioning ``timeout`` must not report an
-    exhausted budget.
-    """
-
-    if _Z3_SOURCE_DIAGNOSTIC.search(message) is not None:
-        return None
-    lowered = message.strip().lower()
-    if "resource limit" in lowered or "canceled" in lowered:
-        return "work"
-    if "memory" in lowered:
-        return "memory"
-    if "timeout" in lowered or "time limit" in lowered:
-        return "time"
-    return None
-
-
-def _project_unknown(reason: str | None) -> tuple[_UnknownResource | None, str]:
-    """Project one Z3 unknown reason onto the typed exhausted-budget taxonomy."""
-
-    text = (reason or "").strip()
-    classified = _classify_exhaustion(text)
-    if classified is not None:
-        return classified, _EXHAUSTION_DETAILS[classified]
-    if not text:
-        return None, "the solver returned no completeness evidence"
-    return None, text[:1_024]
-
-
-def _solver_settings(timeout_ms: int) -> dict[str, int]:
-    """Return the full request-scoped Z3 budget: wall time, work, and memory."""
-
-    return {
-        "timeout": timeout_ms,
-        "rlimit": _SOLVER_RLIMIT,
-        "max_memory": _SOLVER_MAX_MEMORY_MB,
-    }
-
-
-def solve_sat(request: SatSolveRequest) -> SatSolveResult:
-    """Solve one bounded canonical CNF through the maintained Z3 Python binding."""
-
-    try:
-        import z3
-    except (ImportError, OSError) as exc:
-        return SatSolveResult(
-            outcome="UNKNOWN",
-            detail=f"the Z3 backend could not initialize: {exc}"[:1_024],
-        )
-
-    try:
-        variables = tuple(z3.Bool(name) for name in request.cnf.variables)
-        solver = z3.Solver()
-        solver.set(**_solver_settings(request.timeout_ms))
-        for clause in request.cnf.clauses:
-            terms = tuple(
-                variables[abs(literal) - 1]
-                if literal > 0
-                else z3.Not(variables[abs(literal) - 1])
-                for literal in clause
-            )
-            solver.add(z3.Or(*terms))
-        outcome = solver.check()
-        if outcome == z3.sat:
-            model = solver.model()
-            assignment = tuple(
-                z3.is_true(model.eval(variable, model_completion=True))
-                for variable in variables
-            )
-            return SatSolveResult(outcome="SAT", assignment=assignment)
-        if outcome == z3.unsat:
-            return SatSolveResult(outcome="UNSAT")
-    except (OSError, z3.Z3Exception) as exc:
-        exhausted = _classify_exhaustion(str(exc))
-        if exhausted is not None:
-            return SatSolveResult(
-                outcome="UNKNOWN",
-                exhausted=exhausted,
-                detail=_EXHAUSTION_DETAILS[exhausted],
-            )
-        detail = f"the Z3 backend failed during the bounded solve: {exc}"
-        return SatSolveResult(outcome="UNKNOWN", detail=detail[:1_024])
-    exhausted, detail = _project_unknown(solver.reason_unknown())
-    return SatSolveResult(outcome="UNKNOWN", exhausted=exhausted, detail=detail)
-
-
-def _dimacs_cnf(cnf: CanonicalCnf) -> bytes:
-    lines = [f"p cnf {len(cnf.variables)} {len(cnf.clauses)}"]
-    lines.extend(
-        " ".join(str(literal) for literal in clause) + " 0" for clause in cnf.clauses
-    )
-    return ("\n".join(lines) + "\n").encode("ascii")
-
-
-def _ascii_lpr(refutation: SatLprRefutation) -> bytes:
-    lines: list[str] = []
-    for step in refutation.steps:
-        if isinstance(step, LprDeletion):
-            lines.append("0 d " + " ".join(map(str, step.clause_ids)) + " 0")
-            continue
-        fields = [str(step.clause_id), *(str(literal) for literal in step.clause)]
-        if step.witness is not None:
-            fields.extend(str(literal) for literal in step.witness)
-        fields.append("0")
-        fields.extend(str(clause_id) for clause_id in step.at_hint_clause_ids)
-        if not step.propagation_hints:
-            fields.append("0")
-        else:
-            fields.append(str(-step.propagation_hints[0].clause_id))
-            for index, hint in enumerate(step.propagation_hints):
-                fields.extend(str(clause_id) for clause_id in hint.at_hint_clause_ids)
-                fields.append(
-                    str(
-                        -step.propagation_hints[index + 1].clause_id
-                        if index + 1 < len(step.propagation_hints)
-                        else 0
-                    )
-                )
-        lines.append(" ".join(fields))
-    return ("\n".join(lines) + ("\n" if lines else "")).encode("ascii")
-
-
-def _cake_lpr_is_supported(executable: str) -> bool:
-    """Require the source-pinned Cake LPR provider installed by our OCI image."""
-
-    try:
-        return (
-            Path(executable).resolve() == Path("/usr/local/bin/cake_lpr")
-            and _CAKE_LPR_MANIFEST.read_text(encoding="ascii")
-            == _CAKE_LPR_MANIFEST_CONTENT
-        )
-    except OSError:
-        return False
-
-
-def check_sat_refutation(
-    request: SatRefutationCheckRequest,
-) -> SatRefutationCheckResult:
-    """Replay one typed LPR refutation through the pinned CakeML checker."""
-
-    def unavailable(detail: str) -> SatRefutationCheckResult:
-        return SatRefutationCheckResult(
-            outcome="UNAVAILABLE",
-            cnf=request.cnf,
-            refutation=request.refutation,
-            detail=detail,
-        )
-
-    executable = shutil.which("cake_lpr")
-    if executable is None or not _cake_lpr_is_supported(executable):
-        return unavailable(
-            "The source-pinned Cake LPR backend is available only in the Linux service image."
-        )
-    resolved = str(Path(executable).resolve())
-    prlimit = shutil.which("prlimit")
-    if prlimit is not None:
-        prlimit = str(Path(prlimit).resolve())
-    try:
-        with tempfile.TemporaryDirectory(prefix="jacobian-lpr-") as directory:
-            formula_path = Path(directory) / "formula.cnf"
-            proof_path = Path(directory) / "proof.lpr"
-            formula_path.write_bytes(_dimacs_cnf(request.cnf))
-            proof_path.write_bytes(_ascii_lpr(request.refutation))
-            completed = run_bounded_process(
-                [
-                    resolved,
-                    f"--CML_HEAP_SIZE={_LPR_HEAP_MEBIBYTES}",
-                    f"--CML_STACK_SIZE={_LPR_STACK_MEBIBYTES}",
-                    str(formula_path),
-                    str(proof_path),
-                ],
-                input_bytes=b"",
-                timeout_seconds=float(_LPR_WALL_SECONDS),
-                environment=worker_environment(locale="C.UTF-8"),
-                stdout_limit=_LPR_PROCESS_OUTPUT_BYTES,
-                stderr_limit=_LPR_PROCESS_OUTPUT_BYTES,
-                resource_limits=ProcessResourceLimits(
-                    cpu_seconds=_LPR_WALL_SECONDS,
-                    address_space_bytes=_LPR_ADDRESS_SPACE_BYTES,
-                    file_size_bytes=1_024 * 1_024,
-                ),
-                platform_tools=ProcessPlatformTools(prlimit_executable=prlimit),
-                cwd=directory,
-            )
-    except OSError:
-        return unavailable("The fixed Cake LPR backend could not be started.")
-    outcome: Literal["INVALID_REFUTATION", "UNAVAILABLE", "TIMEOUT", "ERROR"]
-    if completed.timed_out:
-        outcome, detail = "TIMEOUT", "Cake LPR exceeded the declared wall-time limit."
-    elif completed.cancelled:
-        outcome, detail = "ERROR", "Cake LPR execution was cancelled."
-    elif completed.stdout_exceeded or completed.stderr_exceeded:
-        outcome, detail = "ERROR", "Cake LPR exceeded the diagnostic-output limit."
-    elif (
-        completed.returncode == 0
-        and completed.stdout == b"s VERIFIED UNSAT\n"
-        and not completed.stderr
-    ):
-        return SatRefutationCheckResult(
-            outcome="VALID_REFUTATION",
-            cnf=request.cnf,
-            refutation=request.refutation,
-        )
-    elif (
-        completed.returncode == 0
-        and not completed.stdout
-        and completed.stderr.startswith(b"c ")
-        and completed.stderr.endswith(b"\n")
-    ):
-        outcome, detail = (
-            "INVALID_REFUTATION",
-            "The typed LPR refutation does not derive contradiction from this CNF.",
-        )
-    else:
-        # Unexpected process output, including resource failures, is never a
-        # mathematical negative verdict.
-        outcome, detail = (
-            "ERROR",
-            "Cake LPR did not produce its exact verified-UNSAT verdict.",
-        )
-    return SatRefutationCheckResult(
-        outcome=outcome,
-        cnf=request.cnf,
-        refutation=request.refutation,
-        detail=detail,
-    )
-
-
-def solve_smt(request: SmtSolveRequest) -> SmtSolveResult:
-    """Solve one bounded SMT-LIB query through the maintained Z3 Python binding."""
-
-    try:
-        import z3
-    except (ImportError, OSError) as exc:
-        return SmtSolveResult(
-            outcome="UNKNOWN",
-            detail=f"the Z3 backend could not initialize: {exc}"[:1_024],
-        )
-
-    try:
-        assertions = z3.parse_smt2_string(request.smtlib)
-        solver = z3.SolverFor(request.logic.value)
-        solver.set(**_solver_settings(request.timeout_ms))
-        solver.add(assertions)
-        outcome = solver.check()
-        if outcome == z3.sat:
-            model = solver.model().sexpr()
-            if len(model.encode("utf-8")) > _MAX_MODEL_BYTES:
-                return SmtSolveResult(
-                    outcome="UNKNOWN",
-                    detail="the satisfying model exceeds the bounded result limit",
-                )
-            return SmtSolveResult(outcome="SAT", model_smtlib=model)
-        if outcome == z3.unsat:
-            return SmtSolveResult(outcome="UNSAT")
-    except (OSError, z3.Z3Exception) as exc:
-        exhausted = _classify_exhaustion(str(exc))
-        if exhausted is not None:
-            return SmtSolveResult(
-                outcome="UNKNOWN",
-                exhausted=exhausted,
-                detail=_EXHAUSTION_DETAILS[exhausted],
-            )
-        detail = f"the Z3 backend failed during the bounded solve: {exc}"
-        return SmtSolveResult(outcome="UNKNOWN", detail=detail[:1_024])
-    exhausted, detail = _project_unknown(solver.reason_unknown())
-    return SmtSolveResult(outcome="UNKNOWN", exhausted=exhausted, detail=detail)
-
-
-LOGIC_OPERATIONS = (
-    MathTool(
-        operation_id="sat.cnf.canonicalize",
-        title="Canonicalize a bounded named CNF",
-        description="Return one canonical CNF; no source, identifier, or artifact is retained.",
-        request_type=CnfCanonicalizeRequest,
-        result_type=CnfCanonicalizeResult,
-        run=canonicalize_cnf,
-        tags=("sat", "cnf", "canonical"),
-        examples=(
-            example(
-                "two_variables",
-                "Normalize a small named CNF.",
-                {"variable_names": ["b", "a"], "clauses": [[1, -2], [2]]},
-            ),
-        ),
-    ),
-    MathTool(
-        operation_id="sat.assignment.check",
-        title="Check a total SAT assignment",
-        description="Evaluate one complete Boolean assignment against one canonical CNF.",
-        request_type=SatAssignmentCheckRequest,
-        result_type=SatAssignmentCheckResult,
-        run=check_sat_assignment,
-        tags=("sat", "cnf", "assignment", "predicate"),
-        examples=(
-            example(
-                "satisfying_assignment",
-                "Check a total assignment against a canonical CNF.",
-                {
-                    "cnf": {"variables": ["a", "b"], "clauses": [[-1, 2], [1]]},
-                    "assignment": [True, True],
-                },
-            ),
-        ),
-    ),
-    MathTool(
-        operation_id="sat.solve",
-        title="Solve a bounded CNF",
-        description="Run the maintained Z3 Python binding on one canonical CNF.",
-        request_type=SatSolveRequest,
-        result_type=SatSolveResult,
-        run=solve_sat,
-        tags=("sat", "cnf", "solve", "z3"),
-        examples=(
-            example(
-                "two_variable_cnf",
-                "Solve a small canonical CNF.",
-                {"cnf": {"variables": ["a", "b"], "clauses": [[-1, 2], [1]]}},
-            ),
-        ),
-    ),
-    MathTool(
-        operation_id="sat.refutation.check",
-        title="Check a bounded LPR SAT refutation",
-        description=(
-            "Replay one typed LPR/ASCII-v1 refutation against its exact canonical "
-            "CNF through the source-pinned CakeML checker. Only VALID_REFUTATION "
-            "establishes UNSAT; unavailable or failed replay is a non-conclusion."
-        ),
-        request_type=SatRefutationCheckRequest,
-        result_type=SatRefutationCheckResult,
-        run=check_sat_refutation,
-        tags=("sat", "cnf", "lpr", "refutation", "certificate"),
-        examples=(
-            example(
-                "unit_contradiction",
-                "Check an LPR empty-clause derivation from two contradictory units.",
-                {
-                    "cnf": {"variables": ["x"], "clauses": [[-1], [1]]},
-                    "refutation": {
-                        "profile": "LPR_ASCII_V1",
-                        "steps": [
-                            {
-                                "kind": "addition",
-                                "clause_id": 3,
-                                "clause": [],
-                                "at_hint_clause_ids": [1, 2],
-                                "propagation_hints": [],
-                            }
-                        ],
-                    },
-                },
-            ),
-        ),
-    ),
-    MathTool(
-        operation_id="smt.solve",
-        title="Solve a bounded SMT-LIB query",
-        description="Run the maintained Z3 Python binding on one QF SMT-LIB query.",
-        request_type=SmtSolveRequest,
-        result_type=SmtSolveResult,
-        run=solve_smt,
-        tags=("smt", "solve", "smtlib", "z3"),
-        examples=(
-            example(
-                "positive_integer",
-                "Solve a bounded quantifier-free linear-integer query.",
-                {
-                    "logic": "QF_LIA",
-                    "smtlib": "(set-logic QF_LIA)\n(declare-const x Int)\n(assert (> x 0))\n(check-sat)",
-                },
-            ),
-        ),
-    ),
-)
-
-__all__ = [
-    "LOGIC_OPERATIONS",
-    "LprAddition",
-    "LprDeletion",
-    "LprPropagationHint",
-    "LprStep",
-    "SatLprRefutation",
-    "SatRefutationCheckRequest",
-    "SatRefutationCheckResult",
-    "SatSolveRequest",
-    "SatSolveResult",
-    "SmtLogic",
-    "SmtSolveRequest",
-    "SmtSolveResult",
-    "check_sat_refutation",
-    "solve_sat",
-    "solve_smt",
-]
