+++ tests/tooling/test_architecture.py
@@ -45,7 +45,7 @@ def test_bounded_process_gateway_requires_external_tool_owner(tmp_path: Path) ->
         "src/jacobian/math/_singular.py",
         source,
     )
-    _write(tmp_path, "src/jacobian/math/logic/_operations.py", source)
+    _write(tmp_path, "src/jacobian/math/logic/_sat.py", source)
     _write(tmp_path, "src/jacobian/math/example/_operations.py", source)
 
     assert _violations(tmp_path, "bounded-process-gateway") == [
