+++ tests/tooling/test_ci_test_plan.py
@@ -51,6 +51,18 @@ def test_canonical_cnf_contract_change_selects_public_contract_evidence() -> Non
     assert plan.run_catalog_examples is True
 
 
+@pytest.mark.parametrize("filename", ["_sat.py", "_smt.py"])
+def test_logic_solver_contract_change_selects_public_contract_evidence(
+    filename: str,
+) -> None:
+    plan = _plan([f"src/jacobian/math/logic/{filename}"])
+
+    assert plan.run_math is True
+    assert plan.math_tests == ("tests/math/logic",)
+    assert plan.run_catalog is True
+    assert plan.run_catalog_examples is True
+
+
 def test_public_operation_kernel_selects_catalog_examples() -> None:
     plan = _plan(["src/jacobian/math/code_theory/_dual_operations.py"])
